requirements.txt
Flask==2.3.3
requests==2.31.0
Flask-WTF==1.2.1
bootstrap-flask==2.3.2
google-cloud-storage==2.13.0
python-dotenv==1.0.0
waitress==2.1.2
werkzeug==3.0.1
flask-session==0.6.0
Flask-SQLAlchemy==3.1.1

# next ones are required for google login
oauthlib==3.0.1
Flask-Login==0.6.3

# next ones are required for database and csv handling
SQLAlchemy==2.0.23
alembic==1.13.0
mysqlclient==2.2.1

multiqc==1.19

celery==5.3.6
redis==5.0.1
psutil==5.9.8

google-api-python-client==2.125.0
boto3>=1.28.57
pyOpenSSL>=23.2.0
cryptography>=41.0.3
pandas

xlrd

openpyxl
debugpy # Required for debugging.
